refactor(animations): separate utils functions
